+++ js/circle.js
@@ -1,0 +1,19 @@
+const Shape = require('./shape.js');
+
+class Circle extends Shape {
+    constructor(fillcolor, backcolor, initials) {
+        super(fillcolor, backcolor, initials);
+    }
+    render() {
+        return `
+            <svg version="1.1" width="300" height="200" xmlns="http://www.w3.org/2000/svg">
+        
+                <circle cx="150" cy="100" r="80" fill="${this.fillcolor}"></circle>
+        
+                <text x="150" y="125" font-size="60" text-anchor="middle" fill="${this.backcolor}">${this.initials}</text>
+        
+            </svg>`;
+    }
+}
+
+module.exports = Circle;
